pyproject.toml
[project]
name = "screw-agents"
version = "0.1.0"
description = "Modular, AI-powered secure code review agents with vulnerability-specific detection knowledge"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Marco (h0pes)" },
]
keywords = ["security", "code-review", "mcp", "sast", "vulnerability-detection"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "mcp",
    "tree-sitter>=0.23",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
screw-agents = "screw_agents.server:main"

[project.urls]
Homepage = "https://github.com/h0pes/screw-agents"
Repository = "https://github.com/h0pes/screw-agents"
Issues = "https://github.com/h0pes/screw-agents/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/screw_agents"]

[tool.hatch.build.targets.wheel.force-include]
"domains" = "screw_agents/domains"

[dependency-groups]
dev = [
    "psycopg[binary]>=3.1",
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "tree-sitter-languages>=1.10",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests", "benchmarks/tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
strict = true
